fix: type 파일 분리
